+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/blockitemsummary/ChainUpdateDetails.java
@@ -11,15 +11,12 @@ import com.concordium.sdk.responses.chainparameters.*;
 import com.concordium.sdk.transactions.CCDAmount;
 import com.concordium.sdk.types.AccountAddress;
 import com.concordium.sdk.types.Timestamp;
-import com.fasterxml.jackson.annotation.JsonUnwrapped;
 import lombok.*;
 
 import java.time.Duration;
-import java.util.Optional;
 
 /**
- * Details of the different chain updates that
- * may occur on the chain.
+ * Details of a successful chain update.
  */
 @EqualsAndHashCode(doNotUseGetters = true)
 @Builder
@@ -157,6 +154,12 @@ public class ChainUpdateDetails {
      */
     private final FinalizationCommitteeParameters finalizationCommitteeParametersUpdate;
 
+    /**
+     * Parameters that govern validator suspension were changed.
+     * This is only non-null if the type is {@link UpdateType#VALIDATOR_SCORE_PARAMETERS}
+     */
+    private final ValidatorScoreParameters validatorScoreParametersUpdate;
+
     public static ChainUpdateDetails from(UpdateDetails update) {
         val chainUpdateDetailsBuilder = ChainUpdateDetails
                 .builder()
@@ -272,6 +275,13 @@ public class ChainUpdateDetails {
                         .type(UpdateType.UPDATE_FINALIZATION_COMMITTEE_PARAMETERS)
                         .finalizationCommitteeParametersUpdate(FinalizationCommitteeParameters.from(payload.getFinalizationCommitteeParametersUpdate()));
                 break;
+            case VALIDATOR_SCORE_PARAMETERS_UPDATE:
+                chainUpdateDetailsBuilder
+                        .type(UpdateType.VALIDATOR_SCORE_PARAMETERS)
+                        .validatorScoreParametersUpdate(ValidatorScoreParameters.from(payload.getValidatorScoreParametersUpdate()));
+                break;
+            case CREATE_PLT_UPDATE:
+                throw new IllegalStateException("This can't happen. CreatePLT operations are not enqueued, but happen immediately");
             case PAYLOAD_NOT_SET:
                 throw new IllegalArgumentException("Unrecognized chain update");