+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/blockitemsummary/Summary.java
@@ -52,6 +52,9 @@ public class Summary {
             case UPDATE:
                 builder.details(Details.newChainUpdate(outcome.getUpdate()));
                 break;
+            case TOKEN_CREATION:
+                builder.details(Details.newTokenCreation(outcome.getTokenCreation()));
+                break;
             case DETAILS_NOT_SET:
                 throw new IllegalArgumentException("Details type is not set.");
         }