+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/TransactionType.java
@@ -5,7 +5,6 @@ import lombok.Getter;
 /**
  * Type of Account Transaction.
  */
-// Types must match https://github.com/Concordium/concordium-base/blob/main/haskell-src/Concordium/Types/Execution.hs
 public enum TransactionType {
     DEPLOY_MODULE((byte) 0),
     INITIALIZE_SMART_CONTRACT_INSTANCE((byte) 1),
@@ -27,7 +26,9 @@ public enum TransactionType {
     ENCRYPTED_TRANSFER_WITH_MEMO((byte) 23),
     TRANSFER_WITH_SCHEDULE_AND_MEMO((byte) 24),
     CONFIGURE_BAKER((byte) 25),
-    CONFIGURE_DELEGATION((byte) 26);
+    CONFIGURE_DELEGATION((byte) 26),
+    TOKEN_UPDATE((byte) 27),
+    ;
 
     /**
      * Number of Bytes used for Serializing {@link TransactionType}.