+++ concordium-sdk/src/main/java/com/concordium/sdk/types/UInt32.java
@@ -1,10 +1,11 @@
 package com.concordium.sdk.types;
 
+import com.concordium.sdk.serializing.CborMapper;
 import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import com.fasterxml.jackson.databind.ser.std.StdSerializer;
+import com.fasterxml.jackson.dataformat.cbor.CBORGenerator;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.val;
@@ -66,25 +67,34 @@ public final class UInt32 implements Comparable<UInt32> {
         return Integer.compareUnsigned(this.value, other.value);
     }
 
+    public UInt32 plus(UInt32 other) {
+        return UInt32.from(this.value + other.value);
+    }
+
     /**
      * A custom Jackson serializer is provided that ensures that the unsigned value
      * is the one used when serializing to JSON.
      */
-    static class UInt32Serializer extends StdSerializer<UInt32> {
-
-        public UInt32Serializer() {
-            this(null);
-        }
-
-        public UInt32Serializer(Class<UInt32> t) {
-            super(t);
-        }
+    static class UInt32Serializer extends JsonSerializer<UInt32> {
 
         @Override
-        public void serialize(
-                UInt32 uint, JsonGenerator jgen, SerializerProvider provider)
-                throws IOException, JsonProcessingException {
-            jgen.writeRawValue(Integer.toUnsignedString(uint.getValue()));
+        public void serialize(UInt32 uint,
+                              JsonGenerator generator,
+                              SerializerProvider provider) throws IOException {
+            if (generator instanceof CBORGenerator) {
+                val cborGenerator = (CBORGenerator) generator;
+
+                // Write as unsigned 4-byte integer.
+                val value = ByteBuffer.allocate(1 + Integer.BYTES)
+                        .put((byte) 0x1A)
+                        .put(uint.getBytes())
+                        .array();
+                CborMapper.writeBytesAsValue(cborGenerator, value, 0, value.length);
+
+                return;
+            }
+
+            generator.writeRawValue(Integer.toUnsignedString(uint.getValue()));
         }
     }
 }